requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
gradio==4.44.0
aiohttp==3.9.1
wikipedia==1.4.0
duckduckgo-search==3.9.6
Pillow==10.1.0
python-dotenv==1.0.0
asyncio==3.4.3
nest-asyncio==1.5.8
requests==2.31.0